Add Apache header for Apache Ant derived SFTP code
